ISSUE-1 refactor maven module
