+++ stream-spark/src/main/scala/com/qiniu/stream/spark/core/JobOperator.scala
@@ -1,12 +1,59 @@
 package com.qiniu.stream.spark.core
 
+import java.io.File
+
 import com.qiniu.stream.spark.core.JobConfig.RichConfig
 import com.qiniu.stream.util.Logging
-import com.typesafe.config.Config
+import com.typesafe.config.{Config, ConfigFactory}
 import org.antlr.v4.runtime.{CharStream, CharStreams}
+import org.apache.spark.sql.SparkSession
 import org.stringtemplate.v4.ST
 
-abstract class JobOperator(config: Config, jobString: Option[CharStream] = None, params: Map[String, String] = Map()) extends Logging {
+import scala.util.{Failure, Success, Try}
+
+/**
+ * Default implementation of Spark Job Operator
+ *
+ */
+case class JobOperator(config: Config,
+                       params: Map[String, String] = Map(),
+                       session: Option[SparkSession] = None,
+                       jobString: Option[CharStream] = None
+                           )   extends Logging {
+
+  private val sparkSession = session.getOrElse(SparkSessionFactory.load(config))
+
+  sys.addShutdownHook(stop())
+
+  def start(): Unit = {
+    val job = Job(jobContent, params)
+    val jobContext = JobContext(job, config)
+
+    def awaitTermination() {
+      if (sparkSession.streams.active.nonEmpty) {
+        if (isDebugEnable || jobContext.isDebugMode) {
+          sparkSession.streams.active.foreach(_.processAllAvailable())
+        } else {
+          sparkSession.streams.awaitAnyTermination()
+        }
+      }
+    }
+
+    import com.qiniu.stream.spark.config.RichStatement._
+    job.statements.foreach(_.execute(jobContext, sparkSession))
+    awaitTermination()
+  }
+
+  def stop(): Unit = {
+    Try {
+      sparkSession.stop()
+    } match {
+      case Success(_) =>
+      case Failure(e) => logError("unexpected error while shutdown", e)
+    }
+  }
+
+
 
   private val jobContent = {
     jobString.getOrElse {
@@ -29,20 +76,37 @@ abstract class JobOperator(config: Config, jobString: Option[CharStream] = None,
     }
   }
 
-  private val job = parseJob(jobContent,params)
+  def isDebugEnable: Boolean = config.isDebugEnable
+}
 
-  sys.addShutdownHook(stop())
+object JobOperator {
 
-  def stop(): Unit = {}
+  def apply(params: Map[String, String]): JobOperator = {
 
-  def start(): Unit = {
-    startJob(job)
+    val config = params.get("config.file") match {
+      case Some(file) => ConfigFactory.parseFile(new File(file))
+      case None => ConfigFactory.load()
+    }
+
+    val jobDsl = params.get("sql.file") match {
+      case Some(file) => Some(CharStreams.fromFileName(file))
+      case None => None
+    }
+
+    val newParams = params - ("config.file", "sql.file")
+
+    JobOperator(config, newParams, None, jobDsl)
+
   }
 
-  protected def parseJob(content: CharStream, params: Map[String, String]): Job
-
-  protected def startJob(job: Job): Unit
-
-  def isDebugEnable: Boolean = config.isDebugEnable
+  def apply(config: Config, session: SparkSession, jobString: String): JobOperator = {
+    require(config != null)
+    require(session != null)
+    require(jobString != null)
+    JobOperator(config, Map(), Some(session), Some(CharStreams.fromString(jobString)))
+  }
 
 }
+
+
+
